+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The plan is to move away from Grunt, move to node:test and node:assert, do the r
 
 ### Changes
 
+- Add .getLocales to allow hacking on selecting locale based on what is available
 - Remove functions that node-gettext already deprecated
 - Add more types
 - Use [optional chaining](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Optional_chaining#browser_compatibility) instead of lodash.get